pyproject.toml
[tool.poetry]
name = "pyadi-hw-api"
version = "0.0.1"
description = "API and logic for embedded hardware management"
authors = ["travis.collins@analog.com"]

[tool.poetry.dependencies]
python = "^3.7.9"
fastapi = "^0.65.0"
uvicorn = "^0.11.1"
pydantic = "^1.3"
requests = "^2.22.0"
loguru = "^0.4.0"
pyadi-jif = {git = "https://github.com/analogdevicesinc/pyadi-jif.git", rev = "main"}
nox = "^2021.6.12"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
autopep8 = "^1.4.4"
ipdb = "^0.12.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
